# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TemplateV4.ApiService | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/auth/confirm-email": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "ConfirmEmail",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/auth/invitations": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "ManageInvitation",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Framework"
        ],
        "operationId": "ListInvitations",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "all"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "sentAt"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfInvitationItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/v1/auth/audit": {
      "get": {
        "tags": [
          "Framework"
        ],
        "operationId": "ListAuditHistory",
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "Action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ActorId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "SubjectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Until",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "at"
            }
          },
          {
            "name": "Direction",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfAuditItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/operations/overview": {
      "get": {
        "tags": [
          "Framework"
        ],
        "operationId": "GetOperationsOverview",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsOverview"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/notifications/summary": {
      "get": {
        "tags": [
          "Framework"
        ],
        "operationId": "GetNotificationSummary",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/notifications": {
      "get": {
        "tags": [
          "Framework"
        ],
        "operationId": "ListNotifications",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPage"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/notifications/read": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "ReadNotifications",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "read",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/notifications/preferences": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "SaveNotificationPreferences",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPreference"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/privacy": {
      "get": {
        "tags": [
          "Framework"
        ],
        "operationId": "GetPrivacyStatus",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/privacy/export": {
      "get": {
        "tags": [
          "Framework"
        ],
        "operationId": "ExportAccountData",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/privacy/email": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "RequestEmailChange",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/privacy/confirm-email": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "ConfirmEmailChange",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/privacy/deletion": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "RequestAccountDeletion",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/privacy/withdraw": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "WithdrawAccountDeletion",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/privacy/requests": {
      "get": {
        "tags": [
          "Framework"
        ],
        "operationId": "ListDeletionRequests",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "requestedAt"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfDeletionItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/privacy/review": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "ReviewAccountDeletion",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDeletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/files": {
      "get": {
        "tags": [
          "Framework"
        ],
        "operationId": "ListFiles",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilePage"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/files/upload": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "UploadFile",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "byte"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/files/{id}/download": {
      "get": {
        "tags": [
          "Framework"
        ],
        "operationId": "DownloadFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/files/{id}/delete": {
      "post": {
        "tags": [
          "Framework"
        ],
        "operationId": "DeleteFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Anonymous-bound antiforgery token returned by GET /api/v1/auth/csrf. The browser must also send an exact allowed Origin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeEmailRequest": {
        "required": [
          "email",
          "proof"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "proof": {
            "$ref": "#/components/schemas/SecurityProof"
          }
        }
      },
      "ConfirmEmailChangeRequest": {
        "required": [
          "challenge"
        ],
        "type": "object",
        "properties": {
          "challenge": {
            "type": "string"
          }
        }
      },
      "ConfirmEmailRequest": {
        "required": [
          "userId",
          "token"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "InvitationRequest": {
        "required": [
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "cancel": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "NotificationPage": {
        "required": [
          "page",
          "unread",
          "optionalEmailEnabled"
        ],
        "type": "object",
        "properties": {
          "page": {
            "$ref": "#/components/schemas/PageOfNotificationItem"
          },
          "unread": {
            "type": "integer",
            "format": "int32"
          },
          "optionalEmailEnabled": {
            "type": "boolean"
          }
        }
      },
      "NotificationPreference": {
        "required": [
          "optionalEmailEnabled"
        ],
        "type": "object",
        "properties": {
          "optionalEmailEnabled": {
            "type": "boolean"
          }
        }
      },
      "NotificationSummary": {
        "required": [
          "unread"
        ],
        "type": "object",
        "properties": {
          "unread": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "OperationsOverview": {
        "required": [
          "pendingMessages",
          "failedMessages",
          "activeJobs",
          "failedJobs",
          "oldestMessageSeconds",
          "lastMaintenanceAt",
          "version",
          "checkedAt",
          "backlogWarningSeconds"
        ],
        "type": "object",
        "properties": {
          "pendingMessages": {
            "type": "integer",
            "format": "int32"
          },
          "failedMessages": {
            "type": "integer",
            "format": "int32"
          },
          "activeJobs": {
            "type": "integer",
            "format": "int32"
          },
          "failedJobs": {
            "type": "integer",
            "format": "int32"
          },
          "oldestMessageSeconds": {
            "type": "number",
            "format": "double"
          },
          "lastMaintenanceAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "version": {
            "type": "string"
          },
          "checkedAt": {
            "type": "string",
            "format": "date-time"
          },
          "backlogWarningSeconds": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageOfAuditItem": {
        "required": [
          "items",
          "total",
          "pageNumber",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageOfDeletionItem": {
        "required": [
          "items",
          "total",
          "pageNumber",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeletionItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PageOfInvitationItem": {
        "required": [
          "items",
          "total",
          "pageNumber",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitationItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PrivacyStatus": {
        "required": [
          "request",
          "deletedFileRetentionDays",
          "notificationRetentionDays"
        ],
        "type": "object",
        "properties": {
          "request": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeletionItem"
              }
            ]
          },
          "deletedFileRetentionDays": {
            "type": "integer",
            "format": "int32"
          },
          "notificationRetentionDays": {
            "type": "integer",
            "format": "int32"
          },
          "reviewPolicy": {
            "type": "string",
            "default": "AdministratorReview"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Use a five-minute access token returned by login. Browser cookie endpoints also require CSRF protection.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}